Added the traslator and an emulator and testing to the project
Added an emulator written in C that allows the user to pipe in a .hack file.
Takes an argument that specifies the number of cycles to run the machine for,
and also can display the screen and take keyboard input.

Added the translator, which converts VM code into hack ASM code.

Added test scripts to the project, including scripts to test the assembler
and translator